# backend/docs/apifox_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Canvas OS API",
    "description": "\n# Life Canvas OS 后端 API\n\n八维生命平衡系统后端接口，支持个人成长管理、日记记录、AI 洞察等功能。\n\n## 主要功能\n\n- **用户认证**: PIN 码设置、验证、修改\n- **系统管理**: 八维系统（饮食、运动、智力、输出、恢复、资产、连接、环境）\n- **用户配置**: 用户信息、设置、AI 配置\n- **日记管理**: 日记 CRUD、附件、编辑历史\n- **AI 洞察**: 基于用户数据生成个性化洞察\n- **数据管理**: 备份、恢复、导出、导入\n\n## 认证方式\n\n本 API 使用 PIN 码认证。\n\n## 响应格式\n\n所有接口遵循统一的响应格式：\n\n```json\n{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1707219200000\n}\n```\n\n## 错误码\n\n- `200`: 成功\n- `201`: 创建成功\n- `400`: 错误请求\n- `401`: 未授权\n- `404`: 资源不存在\n- `409`: 冲突\n- `422`: 参数验证失败\n- `429`: 请求过于频繁\n- `500`: 服务器错误\n\n## 分页参数\n\n| 参数 | 类型 | 必填 | 默认值 | 说明 |\n|------|------|------|--------|------|\n| page | Integer | 否 | 1 | 页码，从 1 开始 |\n| page_size | Integer | 否 | 20 | 每页数量，最大 100 |\n| sort_by | String | 否 | created_at | 排序字段 |\n| sort_order | String | 否 | desc | 排序方向：asc/desc |\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/journal": {
      "post": {
        "tags": [
          "journals",
          "journals"
        ],
        "summary": "创建日记",
        "description": "创建日记",
        "operationId": "create_diary_api_journal_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DiaryResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "journals",
          "journals"
        ],
        "summary": "获取日记列表",
        "description": "获取日记列表",
        "operationId": "get_diaries_api_journal_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "mood",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "great",
                    "good",
                    "neutral",
                    "bad",
                    "terrible"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mood"
            }
          },
          {
            "name": "related_system",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Related System"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "created_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/DiaryResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        },
                        "has_next": {
                          "type": "boolean"
                        },
                        "has_prev": {
                          "type": "boolean"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/journal/{diary_id}": {
      "get": {
        "tags": [
          "journals",
          "journals"
        ],
        "summary": "获取日记详情",
        "description": "获取日记详情",
        "operationId": "get_diary_detail_api_journal__diary_id__get",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/DiaryResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "journals",
          "journals"
        ],
        "summary": "Update Diary",
        "description": "更新日记",
        "operationId": "update_diary_api_journal__diary_id__patch",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiaryUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "journals",
          "journals"
        ],
        "summary": "Delete Diary",
        "description": "删除日记",
        "operationId": "delete_diary_api_journal__diary_id__delete",
        "parameters": [
          {
            "name": "diary_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Diary Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights/generate": {
      "post": {
        "tags": [
          "insights",
          "insights"
        ],
        "summary": "生成洞察",
        "description": "生成洞察",
        "operationId": "generate_insight_api_insights_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InsightGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/InsightGenerateResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/insights": {
      "get": {
        "tags": [
          "insights",
          "insights"
        ],
        "summary": "获取洞察历史",
        "description": "获取洞察历史",
        "operationId": "get_insights_api_insights_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Page Size"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "generated_at",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "desc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/InsightResponse"
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        },
                        "total_pages": {
                          "type": "integer"
                        },
                        "has_next": {
                          "type": "boolean"
                        },
                        "has_prev": {
                          "type": "boolean"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiaryCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "标题"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "title": "Content",
            "description": "内容"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "great",
                  "good",
                  "neutral",
                  "bad",
                  "terrible"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood",
            "description": "心情"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "标签（JSON 数组字符串）"
          },
          "related_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related System",
            "description": "关联系统类型"
          },
          "is_private": {
            "type": "boolean",
            "title": "Is Private",
            "description": "是否私密",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "DiaryCreate",
        "description": "创建日记"
      },
      "DiaryUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "mood": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "great",
                  "good",
                  "neutral",
                  "bad",
                  "terrible"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Mood"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "related_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Related System"
          },
          "is_private": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Private"
          }
        },
        "type": "object",
        "title": "DiaryUpdate",
        "description": "更新日记（所有字段可选）"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InsightGenerateRequest": {
        "properties": {
          "force": {
            "type": "boolean",
            "title": "Force",
            "description": "是否强制重新生成",
            "default": false
          }
        },
        "type": "object",
        "title": "InsightGenerateRequest",
        "description": "生成洞察请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "DiaryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "mood": {
            "type": "string",
            "enum": [
              "great",
              "good",
              "neutral",
              "bad",
              "terrible"
            ]
          },
          "tags": {
            "type": "string"
          },
          "related_system": {
            "type": "string"
          },
          "is_private": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "InsightGenerateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "system_scores": {
            "type": "object"
          },
          "provider_used": {
            "type": "string"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InsightResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "content": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "system_scores": {
            "type": "object"
          },
          "provider_used": {
            "type": "string"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境"
    }
  ],
...
}